engine: refactor
