+++ packages/engine-multi/tsup.config.js
@@ -5,7 +5,7 @@ export default {
   splitting: false,
   entry: {
     index: 'src/index.ts',
-    'worker/worker': 'src/worker/worker.ts',
+    'worker/thread/run': 'src/worker/thread/run.ts',
     // TODO I don't actually want to build this into the dist, I don't think?
     'worker/mock': 'src/worker/mock-worker.ts',
     'worker/child/runner': 'src/worker/child/runner.ts',
